requirements.txt
Package      Version
------------ -------
holmium.core 0.8.5
MarkupSafe   1.1.1
nose         1.3.7
ordereddict  1.1
pip          20.2.1
selenium     3.141.0
setuptools   49.6.0
six          1.15.0
urllib3      1.25.10
wheel        0.35.1
